+++ src/command_modules/azure-cli-sql/azure/cli/command_modules/sql/commands.py
@@ -18,7 +18,8 @@ from ._util import (
     get_sql_firewall_rules_operations,
     get_sql_replication_links_operations,
     get_sql_servers_operations,
-    get_sql_server_usages_operations
+    get_sql_server_usages_operations,
+    get_sql_virtual_network_rules_operations
 )
 
 if not supported_api_version(PROFILE_TYPE, max_api='2017-03-09-profile'):
@@ -214,3 +215,16 @@ if not supported_api_version(PROFILE_TYPE, max_api='2017-03-09-profile'):
             c.generic_update_command('update', 'get', 'create_or_update',
                                      custom_func_name='server_ad_admin_update',
                                      setter_arg_name='properties')
+
+    virtual_network_rules_operations = create_service_adapter(
+        'azure.mgmt.sql.operations.virtual_network_rules_operations',
+        'VirtualNetworkRulesOperations')
+
+    with ServiceGroup(__name__, get_sql_virtual_network_rules_operations, virtual_network_rules_operations,
+                      custom_path) as s:
+        with s.group('sql server vnet-rule') as c:
+            c.command('create', 'create_or_update')
+            c.command('show', 'get')
+            c.command('list', 'list_by_server')
+            c.command('delete', 'delete')
+            c.generic_update_command('update', 'get', 'create_or_update')